+++ Show-PSCalendarHelp.ps1
@@ -1,6 +1,0 @@
-Function Show-PSCalendarHelp {
-    [cmdletbinding()]
-    param()
-
-    Start-process $PSScriptRoot\PSCalendarManual.pdf
-}