+++ functions/Show-PSCalendarHelp.ps1
@@ -1,0 +1,6 @@
+Function Show-PSCalendarHelp {
+    [cmdletbinding()]
+    param()
+    Write-Verbose "Opening $(Resolve-Path $PSScriptRoot\..\PSCalendarManual.pdf)"
+    Start-Process $PSScriptRoot\..\PSCalendarManual.pdf
+}